add report shell command
The report command will produce the output known from `checkin -s` (the
status of the current month).  If there is currently a start/stop cycle
it will report on that as well (current start time and current length).
The same will also be reported if the `start` command is issued more
than once without a corresponding `stop`.